\name{CRC}
\alias{CRC}
\docType{data}
\title{
CRC
}
\description{
A dataset of real microbiome
}
\usage{data(CRC)}
\format{
phyloseq
}

\references{
McMurdie PJ, Holmes S. phyloseq: An R package for reproducible interactive analysis and graphics of microbiome census data. \emph{PLoS ONE}. 2013;8(4):e61217.

Baxter NT, Ruffin MT, Rogers MAM, et al. Microbiota-based model improves the sensitivity of fecal immunochemical test for detecting colonic lesions. \emph{Genome Medicine}. 2016;8(1):37.
}

\examples{
library(phyloseq)

data(CRC)
CRC
}
